fix: adding dependency between tile server test lambda and ecr repo contining test image
