+++ lib/index.ts
@@ -27,5 +27,6 @@ export * from "./osml/osml_vpc";
 export * from "./osml/tile_server/ts_container";
 export * from "./osml/tile_server/ts_dataplane";
 export * from "./osml/tile_server/roles/ts_task_role";
+export * from "./osml/tile_server/testing/ts_test_runner_container";
 export * from "./osml/tile_server/testing/ts_test_runner";
 export * from "./osml/tile_server/testing/ts_imagery";
